+++ src/tasks/instances/goTo.ts
@@ -1,9 +1,9 @@
-import {hasPos} from '../../declarations/typeGuards';
+import {_HasRoomPosition} from '../../declarations/typeGuards';
 import {profile} from '../../profiler/decorator';
 import {Task} from '../Task';
 
 // export type goToTargetType = { pos: RoomPosition } | RoomPosition;
-export type goToTargetType = HasRef & HasPos;  // This is overridden and handled better in the Tasks.goTo() dispatcher
+export type goToTargetType = HasRef & _HasRoomPosition;  // This is overridden and handled better in the Tasks.goTo() dispatcher
 export const goToTaskName = 'goTo';
 
 @profile