+++ src/zerg/AnyZerg.ts
@@ -209,16 +209,15 @@ export abstract class AnyZerg {
 	}
 
 	move(direction: DirectionConstant, force = false) {
-		if (!this.blockMovement && !force) {
-			const result = this.creep.move(direction);
-			if (result == OK) {
-				if (!this.actionLog.move) this.actionLog.move = true;
-				this.nextPos = this.pos.getPositionAtDirection(direction);
-			}
-			return result;
-		} else {
-			return ERR_BUSY;
+		if (this.blockMovement && !force) return ERR_BUSY
+
+		const result = this.creep.move(direction);
+		if (result == OK) {
+			if (!this.actionLog.move) this.actionLog.move = true;
+			this.nextPos = this.pos.getPositionAtDirection(direction);
 		}
+
+		return result;
 	}
 
 	notifyWhenAttacked(enabled: boolean) {
@@ -332,7 +331,7 @@ export abstract class AnyZerg {
 
 	// Movement and location -------------------------------------------------------------------------------------------
 
-	goTo(destination: RoomPosition | HasPos, options: MoveOptions = {}) {
+	goTo(destination: RoomPosition | _HasRoomPosition, options: MoveOptions = {}) {
 		return Movement.goTo(this, destination, options);
 	}
 
@@ -340,7 +339,7 @@ export abstract class AnyZerg {
 		return Movement.goToRoom(this, roomName, options);
 	}
 
-	inSameRoomAs(target: HasPos): boolean {
+	inSameRoomAs(target: _HasRoomPosition): boolean {
 		return this.pos.roomName == target.pos.roomName;
 	}
 
@@ -360,7 +359,7 @@ export abstract class AnyZerg {
 	/**
 	 * Kite around hostiles in the room
 	 */
-	kite(avoidGoals: (RoomPosition | HasPos)[] = this.room.hostiles, options: MoveOptions = {}): number | undefined {
+	kite(avoidGoals: (RoomPosition | _HasRoomPosition)[] = this.room.hostiles, options: MoveOptions = {}): number | undefined {
 		_.defaults(options, {
 			fleeRange: 5
 		});
@@ -368,7 +367,7 @@ export abstract class AnyZerg {
 	}
 
 	private defaultFleeGoals() {
-		let fleeGoals: (RoomPosition | HasPos)[] = [];
+		let fleeGoals: (RoomPosition | _HasRoomPosition)[] = [];
 		fleeGoals = fleeGoals.concat(this.room.hostiles)
 							 .concat(_.filter(this.room.keeperLairs, lair => (lair.ticksToSpawn || Infinity) < 10));
 		return fleeGoals;
@@ -377,7 +376,7 @@ export abstract class AnyZerg {
 	/**
 	 * Flee from hostiles in the room, while not repathing every tick // TODO: take a look at this
 	 */
-	flee(avoidGoals: (RoomPosition | HasPos)[] = this.room.fleeDefaults,
+	flee(avoidGoals: (RoomPosition | _HasRoomPosition)[] = this.room.fleeDefaults,
 		 fleeOptions: FleeOptions              = {},
 		 moveOptions: MoveOptions              = {}): boolean {
 		if (avoidGoals.length == 0 || this.room.dangerousHostiles.find(